Revert "Merge pull request #16 from ishaanp9/ishaan"
This reverts commit 6faa39a24701a3dd363937d864cd3729a5ffd8ee, reversing
changes made to a6f5bc9c17224f67e6ccd2f5ec680eacfe71dcf5.
