+++ AthleteNetwork/App.js
@@ -6,7 +6,6 @@
 //npm install --save react-native-see-more-inline
 //npm add react-native-text-size
 //npm install react-native-elements
-//npm install react-navigation
 
 import React from 'react';
 import { Text, View, SafeAreaView, StyleSheet } from 'react-native';
@@ -14,16 +13,11 @@ import Router from './navigation/router';
 import ChatScreen from './screens/ChatScreen';
 import LoginScreen from './screens/LoginScreensHandler/LoginScreen';
 import TransitionToLoginScreen from './screens/LoginScreensHandler/TransitionToLoginScreen';
-import FeedList from './screens/FeedList';
-import VideoPostItem from './components/VideoPostItem';
-
 
 const App = (props) => {
   return (
     <>
-      {/* <Router/>  */}
-      {/* <ChatScreen/> */}
-      <VideoPostItem/>
+      <Router/> 
     </>
   );
 };
