+++ AthleteNetwork/navigation/bottomTabHandler.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-//import PublicFeedList from '../screens/PublicFeedList';
-import feedTabHandler from './feedTabHandler';
+import FeedList from '../screens/FeedList';
 import ProfileScreen from '../screens/ProfileScreen';
 import AddPostScreen from '../screens/AddPostScreen';
 import ExploreScreen from '../screens/ExploreScreen';
@@ -15,25 +14,15 @@ import Icon from 'react-native-vector-icons/MaterialIcons';
 const Tab = createBottomTabNavigator();
 
 const bottomTabHandler = props => {
-
-
   return (
-    <Tab.Navigator 
-      tabBarOptions={{
-        showLabel: false,
-        // style: {
-        //   //backgroundColor: 'transparent',
-        //   borderTopWidth: 0,
-        //   position: 'absolute',
-        //   left: 0,
-        //   right: 0,
-        //   //height: 50,
-        // }
-      }}
-    >
+    <Tab.Navigator screenOptions={{
+      tabBarLabel: () => {
+        return null;
+      },
+    }}>
       <Tab.Screen
         name={'Feed'}// we can change the name later
-        component={feedTabHandler} 
+        component={FeedList} 
         options={{
           tabBarIcon: ({color}) => (
             <Icon name="home" size={25} color={color} />
