+++ AthleteNetwork/navigation/router.js
@@ -4,11 +4,11 @@ import {createStackNavigator} from '@react-navigation/stack';
 import LoginScreen from '../screens/LoginScreensHandler/LoginScreen';
 import TransitionToLoginScreen from '../screens/LoginScreensHandler/TransitionToLoginScreen';
 
-import PublicFeedList from '../screens/PublicFeedList';
+import FeedList from '../screens/FeedList';
 import bottomTabHandler from './bottomTabHandler';
 import OnboardingSwiper from '../screens/LoginScreensHandler/OnboardingSwiper';
 import ChatScreen from '../screens/ChatScreen';
-// import SignupScreen from '../screens/LoginScreensHandler/SignupScreen';
+import SignupScreen from '../screens/LoginScreensHandler/SignupScreen';
 
 
 const Stack = createStackNavigator();
@@ -16,7 +16,7 @@ const Stack = createStackNavigator();
 const Router = () => {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName={'Onboarding Swiper'}>
+      <Stack.Navigator>
         <Stack.Screen
           name={'Onboarding Swiper'}
           component={OnboardingSwiper}
@@ -31,13 +31,13 @@ const Router = () => {
             headerShown: false,
           }}
         />
-        {/* <Stack.Screen
+        <Stack.Screen
           name={'Signup Screen'}
           component={SignupScreen}
           options={{
             headerShown: false,
           }}
-        /> */}
+        />
         <Stack.Screen
           name={'Login Screen'}
           component={LoginScreen}
@@ -53,8 +53,8 @@ const Router = () => {
           }}
         />
         <Stack.Screen
-          name={'Public Feed List'}
-          component={PublicFeedList}
+          name={'Feed List'}
+          component={FeedList}
           options={{
             headerShown: false,
           }}
