+++ AthleteNetwork/package.json
@@ -13,7 +13,6 @@
     "@eva-design/eva": "^2.1.0",
     "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/bottom-tabs": "^5.11.11",
-    "@react-navigation/material-top-tabs": "^5.3.15",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
     "@ui-kitten/components": "^5.1.0",
@@ -27,12 +26,9 @@
     "react-native-screens": "^3.2.0",
     "react-native-see-more-inline": "^1.3.0",
     "react-native-svg": "^9.13.6",
-    "react-native-tab-view": "^2.16.0",
     "react-native-text-size": "*",
     "react-native-vector-icons": "^8.1.0",
-    "react-native-video": "^5.1.1",
-    "react-navigation": "^4.4.4",
-
+    "react-native-video": "^5.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
