+++ AthleteNetwork/screens/FeedList/styles.js
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+        width: '100%'
+    }
+});
+
+export default styles;
