--- AthleteNetwork/screens/PrivateFeedList/posts.js
@@ -1,38 +1,0 @@
-export default [{
-    id: 12345,
-    username: '@ishaanp9',
-    description: 'We are AthleteNetwork. We connect athletes to different opportunities to make money, such as sponsors, endorsements, and agents.',
-    //descriptionNeedsExpanded: (description.lenth > 15) ? true : false,
-    image: require('../../assets/athletesloginimg.png'),
-    likes: 41,
-    comments: 2,
-    reposts: 0,
-    
-  }, {
-    id: 12346,
-    username: '@AthleteX',
-    description: 'We are AthleteNetwork.',
-    image: require('../../assets/thxraghav.png'),
-    likes: 39,
-    comments: 3,
-    reposts: 1,
-  }, {
-    id: 12347,
-    username: '@ethank.6',
-    description: 'Yo',
-    image: require('../../assets/athletesloginimg.png'),
-    likes: 12,
-    comments: 2,
-    reposts: 1,
-  }, {
-    id: 12348,
-    username: '@athletenetwork',
-    description: 'We are AthleteNetwork. We connect athletes to different opportunities to make money, such as sponsors, endorsements, and agents.',
-    image: require('../../assets/thxraghav.png'),
-    likes: 14,
-    comments: 1,
-    reposts: 1,
-    //tagline: 'Order Online for',
-    //taglineCTA: 'Touchless Delivery',
-    //image: require('../../assets/images/ModelY.jpeg'),
-  }];
