--- AthleteNetwork/screens/PrivateFeedList/styles.js
@@ -1,9 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-    container: {
-        width: '100%'
-    }
-});
-
-export default styles;
